requirements.txt
# Gemini packages
google-generativeai>=0.3.0
langchain-google-genai>=0.0.2

# Vector store and embeddings
faiss-cpu>=1.7.4
langchain>=0.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Testing
pytest>=7.4.0